+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed 
 - `MRI/Nifti_SPM12`: the `JSONDecodeError` is no more cilenced in `__isValidFile` method
 
+### Added
+- plugin option to `bidsmapper`
+
 ## [dev2.0.2] - 2020-01-27
 
 ### Added